solutions/requirements.txt
pwntools
ipython
